# scm-service/api/openapi/scm-api.yaml
openapi: 3.0.3
info:
  title: SCM Security & Compliance API
  description: API для управления безопасностью и соответствием требованиям в логистической экосистеме Nodeorb
  version: 1.0.0
  contact:
    name: Nodeorb Team
    email: support@nodeorb.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://scm.nodeorb.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ValidateManualInputRequest:
      type: object
      required:
        - user_id
        - order_id
        - materials_cost
        - labor_cost
        - currency
      properties:
        user_id:
          type: string
          description: Идентификатор пользователя
        order_id:
          type: string
          description: Идентификатор заказа
        service_source:
          type: string
          default: "freight-marketplace"
          description: Источник запроса
        materials_cost:
          type: number
          format: double
          description: Стоимость материалов
        labor_cost:
          type: number
          format: double
          description: Стоимость работ
        currency:
          type: string
          default: "USD"
          description: Валюта
    ValidateManualInputResponse:
      type: object
      properties:
        status:
          type: string
          enum: [GREEN, YELLOW, RED]
          description: Статус валидации
        risk_score:
          type: number
          format: double
          description: Оценка риска (0.0 - 1.0)
        requires_appeal:
          type: boolean
          description: Требуется апелляция
        requires_biometrics:
          type: boolean
          description: Требуется биометрическая проверка
        suggested_median:
          type: number
          format: double
          description: Предложенная медианная цена
        comment:
          type: string
          description: Комментарий к решению
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
paths:
  /api/v1/auth/sync-keycloak:
    post:
      summary: Синхронизация данных пользователя из Keycloak
      description: Вызывается Keycloak при регистрации или изменении профиля пользователя
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                email:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Успешная синхронизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
  /api/v1/policies/manual-input-validation:
    post:
      summary: Валидация ручного ввода стоимости
      description: Проверка введенных цен на материалы и работы через Market Oracle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateManualInputRequest'
      responses:
        '200':
          description: Результат валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateManualInputResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...